+++ semcore/time-picker/src/PickerInput.jsx
@@ -124,23 +124,11 @@ class ItemPicker extends Component {
           neighborLocation={false}
           value={value}
           aria-label={this.getAriaLabel()}
-          role='combobox'
-          aria-autocomplete='list'
-          aria-expanded={visible ? 'true' : 'false'}
-          aria-haspopup='listbox'
-          aria-controls={`picker-${this.field}-popper`}
           onChange={this.handleChange}
           onBlur={this.handleBlur}
           onKeyDown={this.handleKeyDown}
         />
-        <Select.Menu
-          aria-hidden='true'
-          hMax={180}
-          role='listbox'
-          id={`picker-${this.field}-popper`}
-        >
-          {getOptions(min, max, step)}
-        </Select.Menu>
+        <Select.Menu hMax={180}>{getOptions(min, max, step)}</Select.Menu>
       </Select>,
     );
   }
@@ -231,15 +219,6 @@ class Minutes extends ItemPicker {
       }
     }
     if (event.key === 'Enter') this.submitChanges(event);
-    if (event.key === 'ArrowRight') {
-      if (
-        currentTarget.selectionStart >= currentTarget.value.length &&
-        currentTarget.selectionStart === currentTarget.selectionEnd
-      ) {
-        event.preventDefault();
-        this.focusNext();
-      }
-    }
   };
 }